List classes and tests bugfixes. Default `activewhen` rule.
MenuItem's now render their own `li`'s which also includes the activeness class. One can also pass additional classes to the `li`'s.
Provide a default rule for `activewhen` which just matches the request endpoint to the menu item's endpoint.
Only raise `RuntimeError`'s when the flask application is in debug mode, if not, raise `RuntimeWarning`'s.
`MenuItemContent` now accepts a callable for it's content allowing the content code to be generated when rendering.
Fix tests to include the changes added.